Removes non-functional meson debug option.
It looks like the debug mode was copied literally from the old
Makefiles, which interacts poorly with meson and leads to builds
with "-O0 -g -O3".

The correct way to do this is to use "-Dbuildtype=release" or
"-Dbuildtype=debug" during configuration. Set the default builds
made by scripts/build to release.