Implement argon2.
Fully configurable as per the spec. Includes a new type, in addition to
i, d, and id: u. Similar to id in that it is a hybrid solution, but
slightly tweaked to be unique to Urbit's implementation.

Initially implemented as a direct port from the pure Python
implementation, then lightly rewritten to be more Hoon-ish.
https://pypi.org/project/argon2pure/