fix core nest mismatch
There is a variance bug such that the following cast is allowed to go through:

```
> =a 12

> =core1 ^?  |=(b=@ a)

> =a "hello"

> =core2 ^?  |=(b=@ a)

> ^+(core1 core2)
```

This should not be!  The code in `hoon.hoon` is correct -- it turns out that the jet is wrong however.  It falsely returns "yes" for a nest in which two cores have identical code.  Obviously that's not enough to guarantee sameness of type.  So I've removed the offending lines.