claz: Implement batch invites
For generating many sendPoint() transactions for the Delegated Sending
contract. Specify what ship to send the invites as, and a path to a file
containing lines of "~ship,~ticket,0xaddress".

Comes with a generator, |claz-invites, for generating such files, given
a star and a range of its children (and an output path).