vere: change default bootstrap pill URI
Vere will attempt to download a pill from https://bootstrap.urbit.org if
not given one explicitly.  These pill filenames have traditionally had
the format:

  urbit-$URBIT_VERSION.pill

but this is inconsistent with the manner in which versions are referred
to elsewhere.  For example, release binaries are packaged in tarballs
with the names:

  urbit-linux64-v$URBIT_VERSION.tgz
  urbit-darwin-v$URBIT_VERSION.tgz

Note the 'v' prepended to URBIT_VERSION.

From v0.9.2 forward it is expected that bootstrap pills will also use
the 'v' prefix, i.e.

  urbit-v$URBIT_VERSION.pill

so this commit makes the appropriate change in the daemon.