sh: add OTA test helper script [ci skip]
Adds a script for helping to verify that an OTA update (i.e. from arvo
at one revision to another) applies cleanly.

This only performs the upgrade on a single fake ship, so it's limited in
what kind of problems it can identify.  It *can* catch particularly
nasty errors, however, so serves as a useful sanity check.