soto: mutate cursor for tab complete inserts
This changes sole.js to handle cursor behaviour for tabbing.

It does so by overriding the cursor position in transpose if it's greater
than it expects. This could produce errant results if other sole apps
use insert behind text, but that seems like an edge case. Flagged for
future with a comment.