chat fe: invite search with/out ships
This lets you specify whether or not you want to include ships in search
results for the InviteSearch component, as is already possible for
groups. This enables group-only searching, as will be used by the next
commit.

Also modifies the placeholder text based on what is included in the
search results.