shoe: add %table and %row shoe commands
For rendering a data table or just a single table row respectively.

Table data is a list of header names, a list of width restrictions, and a list
of rows, defined as lists of dimes.
Row contents are rendered as per the aura in the dime. This also determines
their alignment (left/right), and how they break to fit smaller-width columns.

%row was added because %table necessitated implementing it. Whether it's a good
fit for the shoe "protocol" remains to be seen.