landscape: fix initial contact color input
The rewrite of hexToUx in TypeScript returned a function script instead
of the expected string.

The previous write is put in its place, and instead of assuming "0x"
is always in the color input's initial value, we selectively remove it,
alleviating another bug where we would display a different color than we
submitted due to substringing the first 2 characters
of a value without "0x".

Fixes #3875.