term: restore cursor position after draw calls
This keeps the cursor position at the last %hop location. And being able
to go back to that position without relying on state is just plain
convenient.

The exception here is %nel blits, which move the cursor (down and) to
the start of the line, as expected of "new line" semantics.