vere: request binary mode in fopen/open calls
The C runtime MingW relies on, msvcrt.dll, has the notion of accessing files in text and binary mode.
In text mode, it does CRLF conversion and handles some control characters, which breaks binary files.
This change adds O_BINARY in every fopen/open call and #defines it to 0 on other platforms.