vere: basic MingW compatibility changes
This commit adds code changes, compatibility functions, stubs and a build script
to build urbit binaries on MingW64. Some functionality is limited or missing:
terminal input and daemon mode is not available, graceful exit does not work,
and the binaries are not completely static and use (portable) MingW dlls.

To build the binaries, install the MSYS2 environment, check out or copy the urbit
repo and pill binaries, open a MingW64 shell and `cd pkg/urbit && ./build-mingw`.