kh: decouple trace & slog writing from drum prompt
New blit logic already stopped assuming the bottom line based on %nel
blits, instead looking at cursor position to determine if something
was getting drawn to the bottom of the screen or not.

Here, we stop overwriting the bottom line entirely (except for the
spinner), instead inserting the trace/slog directly above the bottom
line on-screen, without overwriting anything.

Side-effect of this is that trailing newlines are always there, so we
can stop including them explicitly.