interface: catch contact-update type mismatch
When applying add-group or remove-group updates optimistically, we send
our ship resourceFromPath objects, and get back /ship/~name/resource
strings. When we apply the update optimistically, we use the former,
storing the object in state. This causes interface errors, as it expects
the string. Therefore, on both ends we construct the string and dedupe
if it comes in after application.

Fixes urbit/landscape#917