docket: improve upload page
Most notably, we further restrict the desks eligible for glob upload
to just the ones whose glob is said to be hosted on the local ship.

On the upload page, we replace the desk input field with a dropdown
from which eligible desks can be selected.
Additionally we now display some instructions, instead of leaving
developers entirely on their own.