multipart: more lenient content-type parsing
It's not uncommon for files to have mime types like `image/svg+xml`.
The + in there was tripping up the strictly-`@ta` parser for the content-type.

Here we change the parser to accept +, and all the other characters that the
`@t` parser accepts. This is perhaps a bit overkill, but we primarily care
about getting the parsed request to the caller. If any further validation is
required, it can happen upstream.

Fixes #5282.