helm: add poke to helm to allow self-breaching moons
Intended use is for transitory moons to be able to breach themselves on
startup.

If you run a moon without persistence, then every time the program is
restarted, it must be breached.  This lets the moon breach itself
instead of requiring direct interaction with the planet.  The moon
should reserve the first bone for this purpose, and then every time it
starts up, it should send [%helm-moon-breach ~moon-name] to hood on the
planet.