http: check requests for auth using eyre's tokens
Accept the newly added %session effect. Store the tokens received in it,
and refer back to them when checking requests for authentication.

Adds logic for checking requests for authentication based on the
presence of a valid urbauth cookie for the host ship, and updated the
slogstream endpoint to use that instead of the scry-based authentication
check.