ci: remove the mingw build flow
As discussed internally at the end of last year, we are discontinuing
Windows support for vere.

The tldr is that we presently don't have the resources to support
non-*nix systems at the level of quality we want to deliver. Notably
Windows is already very much a second-class citizen (#5822 and others).
Windows users may want to run Urbit through WSL instead.

As such, the Window build step should be removed from our CI.