tomb: remove |rm prompt
if you're trying to tombstone at the head of the desk, you probably
don't know what you're doing. so we abort.

we keep the option to `|rm` any matching hashes in other desks since
this is something the tombstoner might not know exists in advance and is
actively blocking them from completing the desired tombstone operation.