eyre: dynamic responses for empty cache entries
If the cache entry for any given url was cleared, eyre would recognize
that an entry was bound at some point in the past, and serve a 404,
instead of falling back to the normal, cacheless, dynamic response.

Here, we make eyre serve from the cache only if there is a full cache
entry for the url, and have it proceed into normal request handling
logic otherwise.

This way, setting a cached response for some url no longer means giving
up dynamic responses on that url for the rest of time.

Recent versions of vere exhibit the same behavior. As such, in practice,
this code won't get hit without a similarly-patched version of the
runtime.