gall: respond to /whey scries (for |mass)
The restructuring in #6852 put everything under versioned scry paths.
Arvo remains none the wiser and still happily scries for /whey, without
a leading version number, resulting in loss of detail for gall's section
in |mass output.

Here, we simply slap the current version number in front of the /whey
scry path, to get it to resolve.