mesa: if we have nax state for a message, produce it
Previously we were no-oping if the message was older than the
migration line but this could mean that (n)acks that got lost for
could make the flow get stuck.

The solution is to always check if we have state for naxplanation
messages and if so produce the ack. In the migration we were already
binding the live naxplanation on our namespace so we need to look
there first.

In the case of %acks, can we guarantee that line.state was an ack?

In theory we can't guarantee it just by looking at the state of the flow,
but, if it was a %nack we would have state in nax.state for live
naxplanations and if the naxplanation had suceeded then they are
not going to resend the payload anymore.

if line.state was an %ack but it got lost we can not know for sure, but,
because we were not removing the correct message from nax.sink
it's very likely that if line.state is not in nax.state that's because it
was indeed a %nack.