eyre: separate session timeout for guest and auth
Presently, eyre creates a new guest session for every unauthenticated
request it gets. For ships accessible from the clearweb, crawlers and
other bots cause this to happen often. Sessions time out and get cleared
up by eyre after some period of inactivity. As such, the amount of
stored (guest) sessions is soft-bounded at "the amount of
unauthenticated requests received within [session timeout time]".

To keep normal users from having to log in after a short time away from
the keyboard, we want to increase the session timeout time. But this
would loosen the bound described above.

Here, we separate the timeout times for guest sessions and "real",
authenticated sessions. This lets us tweak their timeouts independently.
We immediately use this to increase the time for real sessions to 30
days, but leave guest sessions untouched for now.

(Eyre doesn't strictly need to create a whole guest session if the
incoming request wouldn't hit userspace. `+request` already has a TODO
for a refactoring that would unlock that efficiency. But we punt on it
for now. The change here gets us 80% of the desired effect, anyway.)