hoon: make +rest take unary arg instead of list
+rest accepted a list of type-hoon pairs, but was only ever called with
a single item (in +repo).

Here, we update it and its callsite to only accept and pass the single
argument. This lets us simplify and modernize the code slightly, and
avoids some call overhead.

In practice, the performance gains from this are negligible. It may be
_very slightly_ faster, certainly not any slower. Even without
performance gains, the legibility gains here seem worth the change.

Co-authored-by: mopfel-winrux <amadeo.bellotti@gmail.com>