mesa: handle old bone wires with no rift
This was a leftover from the times when the |mesa core was only handling %mesa related wires.

In this case, all wires have a rift, and this is the same in |ames, but previously they didn't, until the fix for dangling-bones introduced it—this was related to subscription wires for old rifts (i.e. after the peer breached).

|mesa needs to handle this as well, since migrated peers can still have this wires (e.g. if the peer has never breached and the subscription is old enough that doesn't have the rift in it).

(This can be a serious issue, if triggered, for anybody that has |ahoyed another ship, since it has caused at least one instance of a peer losing state so i'm targeting master so we can release this before the other mesa bug fixes that need a bit more testing)