eyre: fix double boot protection bone format
The runtime makes the http requests for double boot protection using the standard C `%d` format but the endpoint in eyre expects the bone to be `@ud` formatted. This causes a mismatch if the ping-bone is larger than 1000, causing double boot protection to become disabled. We address the mismatch in arvo instead of vere because this is easier to release.